requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
google-generativeai
pdfkit
pillow
wordcloud
